code-refactoring, Joined all offsite Command Pool classes into one common for all offsite payments, also created single class for all Offsite Initialize Command Response Validators also renamed classes of Value Handlers from Credit Card Payment method to be more accurate for Credit Card, added more comments
